+++ packages/blocks-core/js/libs/wordpress/accordion-heading/test/functionality.blocks-1.e2e.cy.js
@@ -50,6 +50,27 @@ describe('Accordion Heading Block', () => {
 
 		cy.checkBlockCardItems(['normal', 'hover', 'elements/icon']);
 
+		cy.checkBlockStatesPickerItems(
+			[
+				'states/hover',
+				'states/focus',
+				'states/focus-within',
+				'states/first-child',
+				'states/last-child',
+				'states/only-child',
+				'states/empty',
+				'states/before',
+				'states/after',
+				'elements/bold',
+				'elements/italic',
+				'elements/kbd',
+				'elements/code',
+				'elements/span',
+				'elements/mark',
+			],
+			true
+		);
+
 		//
 		// 1. Edit Block
 		//